Eliminate api project
