+++ build.gradle.kts
@@ -39,9 +39,6 @@ spotless {
 }
 
 dependencies {
-  compileOnly(localGroovy())
-  api(projects.api)
-
   implementation("org.jdom:jdom2:2.0.6.1")
   implementation("org.ow2.asm:asm-commons:9.7.1")
   implementation("commons-io:commons-io:2.17.0")
