+++ settings.gradle.kts
@@ -30,6 +30,3 @@ dependencyResolutionManagement {
 rootProject.name = "shadow"
 
 enableFeaturePreview("STABLE_CONFIGURATION_CACHE")
-enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
-
-include(":api")
