+++ src/main/kotlin/com/github/jengelman/gradle/plugins/shadow/ShadowBasePlugin.kt
