+++ src/main/kotlin/com/github/jengelman/gradle/plugins/shadow/ShadowPlugin.kt
