+++ src/main/kotlin/com/github/jengelman/gradle/plugins/shadow/internal/CleanProperties.kt
