+++ src/main/kotlin/com/github/jengelman/gradle/plugins/shadow/tasks/InheritManifest.kt
