+++ src/main/kotlin/com/github/jengelman/gradle/plugins/shadow/tasks/KnowsTask.kt
