+++ src/main/kotlin/com/github/jengelman/gradle/plugins/shadow/transformers/ApacheLicenseResourceTransformer.kt
@@ -28,14 +28,14 @@ import org.gradle.api.file.FileTreeElement
  * @author John Engelman
  */
 public class ApacheLicenseResourceTransformer : Transformer by NoOpTransformer {
-    override fun canTransformResource(element: FileTreeElement): Boolean {
-        val path = element.relativePath.pathString
-        return LICENSE_PATH.equals(path, ignoreCase = true) ||
-          LICENSE_TXT_PATH.regionMatches(0, path, 0, LICENSE_TXT_PATH.length, ignoreCase = true)
-    }
+  override fun canTransformResource(element: FileTreeElement): Boolean {
+    val path = element.relativePath.pathString
+    return LICENSE_PATH.equals(path, ignoreCase = true) ||
+      LICENSE_TXT_PATH.regionMatches(0, path, 0, LICENSE_TXT_PATH.length, ignoreCase = true)
+  }
 
-    public companion object {
-        private const val LICENSE_PATH = "META-INF/LICENSE"
-        private const val LICENSE_TXT_PATH = "META-INF/LICENSE.txt"
-    }
+  public companion object {
+    private const val LICENSE_PATH = "META-INF/LICENSE"
+    private const val LICENSE_TXT_PATH = "META-INF/LICENSE.txt"
+  }
 }
