+++ src/main/kotlin/com/github/jengelman/gradle/plugins/shadow/transformers/Transformer.kt
