+++ src/main/kotlin/com/github/jengelman/gradle/plugins/shadow/transformers/TransformerContext.kt
